Move metada in separate module
